Inline base classes now that there is only one backend (#650)
* inline abstract class

* remove redundant code

* Revert "remove redundant code"

This reverts commit fc3df513319f1156494d68435d4af79c9b4d1daa.

* reremove old base
